+++ plone/app/users/tests/test_user_data_panel.py
@@ -1,5 +1,6 @@
 from plone.app.users.browser.personalpreferences import UserDataPanel
 from plone.app.users.tests.base import TestCase
+from zExceptions import NotFound
 from zope.i18n import translate
 
 
@@ -14,6 +15,8 @@ class TestUserDataPanel(TestCase):
         form = UserDataPanel(portal, request)
         description = translate(form.description, context=request)
         self.assertTrue('admin' in description)
+        # form can be called without raising exception.
+        self.assertTrue(form())
 
     def test_escape_html(self):
         portal = self.portal
@@ -24,3 +27,4 @@ class TestUserDataPanel(TestCase):
         form = UserDataPanel(portal, request)
         description = translate(form.description, context=request)
         self.assertTrue('<script>' not in description)
+        self.assertRaises(NotFound, form)